ready for review
stylist dropdowns should be integrated into appointment/unavailability forms. Need to click arrow to get full view of dropdown caused by wrapping the mat-option elements in a mat-select tag. onSelectionChange event function was associated with the option rather than the select and was passed a stylist rather than an event